requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.4
torch==2.1.0
torchvision==0.16.0
realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
opencv-python==4.8.1.78
pydantic==2.5.0